requirements.txt
pillow==10.4.0

asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.1
